fix(match): move mergeNewFreezesOntoFresh out of "use server" module
A \"use server\" file may only export async Server Actions, so exporting
the synchronous pure helper from publishRoundSummary.ts broke the
Turbopack production build (\"Server Actions must be async functions\").
Move it to lib/match/frozenTileMerge.ts (pure logic belongs in lib
anyway) and import it into the action + test. No behaviour change.
